% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_grm.R
\name{gen_dmeasure_grm}
\alias{gen_dmeasure_grm}
\title{Generate grm dmeasure C code}
\usage{
gen_dmeasure_grm(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\value{
A Csnippet
}
\description{
Generate grm dmeasure C code
}
