% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_lin.R
\name{gen_dmeasure_lin}
\alias{gen_dmeasure_lin}
\title{Generate linear dmeasure C code}
\usage{
gen_dmeasure_lin(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\value{
A Csnippet
}
\description{
Generate linear dmeasure C code
}
