% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_states_lin.R
\name{gen_rinit_lin}
\alias{gen_rinit_lin}
\title{Generate linear rinit C code}
\usage{
gen_rinit_lin(process_args)
}
\arguments{
\item{process_args}{Named list of process arguments}
}
\value{
A Csnippet
}
\description{
Generate linear rinit C code
}
