% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_grm.R
\name{gen_rmeasure_grm}
\alias{gen_rmeasure_grm}
\title{Generate grm rmeasure C code}
\usage{
gen_rmeasure_grm(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\value{
A Csnippet
}
\description{
Generate grm rmeasure C code
}
