% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_lin.R
\name{gen_rmeasure_lin}
\alias{gen_rmeasure_lin}
\title{Generate linear rmeasure C code}
\usage{
gen_rmeasure_lin(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\value{
A Csnippet
}
\description{
Generate linear rmeasure C code
}
