% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_states_lin.R
\name{gen_rprocess_lin}
\alias{gen_rprocess_lin}
\title{Generate linear rprocess C code}
\usage{
gen_rprocess_lin(process_args)
}
\arguments{
\item{process_args}{Named list of process arguments}
}
\value{
A Csnippet
}
\description{
Generate linear rprocess C code
}
