% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_grm.R
\name{get_transmeasure_grm}
\alias{get_transmeasure_grm}
\title{Transform grm parameters}
\usage{
get_transmeasure_grm(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\value{
A named list containing the parameters
}
\description{
Transform grm parameters
}
