% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_lin.R
\name{get_transmeasure_lin}
\alias{get_transmeasure_lin}
\title{Transform linear parameters}
\usage{
get_transmeasure_lin(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\value{
A named list containing the parameters
}
\description{
Transform linear parameters
}
