% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_states_lin.R
\name{get_transprocess_lin}
\alias{get_transprocess_lin}
\title{Transform linear process parameters}
\usage{
get_transprocess_lin(process_args)
}
\arguments{
\item{process_args}{Named list of process arguments}
}
\value{
A named list containing the parameters
}
\description{
Transform linear process parameters
}
