% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statespace_model.R
\name{lin_two_state_statespacemodel}
\alias{lin_two_state_statespacemodel}
\title{Generate a two-state linear measure state-space model object.}
\usage{
lin_two_state_statespacemodel(
  data = NULL,
  n_measures,
  process_mat,
  loadings_mat,
  sigmas,
  times = NULL,
  n_particles = 500,
  n_mif2_iter = 1000,
  sim = FALSE
)
}
\arguments{
\item{data}{The provided dataset. Defaults to NULL.}

\item{n_measures}{The number of measures.}

\item{process_mat}{A 2 x 2 transition matrix for each state.}

\item{loadings_mat}{A (n_measure x 2) matrix of state-measure loadings. Cross-loadings allowed.}

\item{sigmas}{A vector defining standard deviations for each measure.}

\item{times}{A vector sequence of times.}

\item{n_particles}{Number of particles used to estimate parameters. Defaults to 500.}

\item{n_mif2_iter}{Number of iterations on a mif2 run. Defaults to 1000.}

\item{sim}{Whether or not a model should be fit or data should be simulated. If FALSE, returns a pomp object. Defaults to FALSE.}
}
\value{
An S4 statespace_mod class or a pomp object.
}
\description{
Generate a two-state linear measure state-space model object.
}
