% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statespace_model.R
\docType{class}
\name{statespace_mod-class}
\alias{statespace_mod-class}
\title{An S4 class for fitting and validating state-space models}
\description{
An S4 class for fitting and validating state-space models
}
\section{Slots}{

\describe{
\item{\code{pomp_obj}}{pomp.}

\item{\code{data}}{data.frame.}

\item{\code{measure_type}}{character. List of named parameters for the r and d measure functions.}

\item{\code{process_type}}{character. List of named parameters for the r and d process functions.}

\item{\code{measure_args}}{list .}

\item{\code{process_args}}{list.}

\item{\code{time_args}}{list.}

\item{\code{mif2_args}}{list.}

\item{\code{param_traces}}{list.}

\item{\code{filtered_states}}{list.}

\item{\code{mif2_diagnostics}}{data.frame.}
}}

