% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statespace_model.R
\name{statespace_model}
\alias{statespace_model}
\title{Generate a state-space model object}
\usage{
statespace_model(
  data = NULL,
  n_states,
  n_measures,
  measure_type,
  process_type = "lin",
  measure_args,
  process_args,
  dt = NULL,
  times = NULL,
  n_particles = 500,
  n_mif2_iter = 1000,
  sim = FALSE
)
}
\arguments{
\item{data}{The provided dataset. Defaults to NULL.}

\item{n_states}{The number of states.}

\item{n_measures}{The number of measures.}

\item{measure_type}{Measure type selected ("grm" or "lin").}

\item{process_type}{Process type selected. Defaults to "lin".}

\item{measure_args}{List of named parameters for the r and d measure functions.}

\item{process_args}{List of named parameters for the r and d process functions.}

\item{dt}{Step size for Euler's method. If NULL, discrete time. Defaults to NULL.}

\item{times}{Vector sequence of times.}

\item{n_particles}{Number of particles used to estimate parameters. Defaults to 500.}

\item{n_mif2_iter}{Number of iterations on a mif2 run. Defaults to 1000.}

\item{sim}{Whether or not a model should be fit or data should be simulated. If FALSE, returns a pomp object. Defaults to FALSE.}
}
\value{
An S4 statespace_mod class or a pomp object.
}
\description{
Generate a state-space model object
}
