% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_grm.R
\name{validate_measure_args_grm}
\alias{validate_measure_args_grm}
\title{Validate grm measure arguments}
\usage{
validate_measure_args_grm(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\description{
Validate grm measure arguments
}
