% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_meas_lin.R
\name{validate_measure_args_lin}
\alias{validate_measure_args_lin}
\title{Validate linear measure arguments}
\usage{
validate_measure_args_lin(measure_args)
}
\arguments{
\item{measure_args}{Named list of measure arguments}
}
\description{
Validate linear measure arguments
}
