% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_states_lin.R
\name{validate_process_args_lin}
\alias{validate_process_args_lin}
\title{Validate linear process arguments}
\usage{
validate_process_args_lin(process_args)
}
\arguments{
\item{process_args}{Named list of process arguments}
}
\description{
Validate linear process arguments
}
